+++ 3rdparty/glsl-optimizer/src/glsl/lower_mat_op_to_vec.cpp
@@ -126,14 +126,14 @@ ir_mat_op_to_vec_visitor::do_mul_mat_mat(ir_dereference *result,
    ir_assignment *assign;
    ir_expression *expr;
 
-   for (b_col = 0; b_col < b->type->matrix_columns; b_col++) {
+   for (b_col = 0; b_col < (int)b->type->matrix_columns; b_col++) {
       /* first column */
       expr = new(mem_ctx) ir_expression(ir_binop_mul,
 					get_column(a, 0),
 					get_element(b, b_col, 0));
 
       /* following columns */
-      for (i = 1; i < a->type->matrix_columns; i++) {
+      for (i = 1; i < (int)a->type->matrix_columns; i++) {
 	 ir_expression *mul_expr;
 
 	 mul_expr = new(mem_ctx) ir_expression(ir_binop_mul,
@@ -164,7 +164,7 @@ ir_mat_op_to_vec_visitor::do_mul_mat_vec(ir_dereference *result,
 				     get_element(b, 0, 0));
 
    /* following columns */
-   for (i = 1; i < a->type->matrix_columns; i++) {
+   for (i = 1; i < (int)a->type->matrix_columns; i++) {
       ir_expression *mul_expr;
 
       mul_expr = new(mem_ctx) ir_expression(ir_binop_mul,
@@ -185,7 +185,7 @@ ir_mat_op_to_vec_visitor::do_mul_vec_mat(ir_dereference *result,
 {
    int i;
 
-   for (i = 0; i < b->type->matrix_columns; i++) {
+   for (i = 0; i < (int)b->type->matrix_columns; i++) {
       ir_rvalue *column_result;
       ir_expression *column_expr;
       ir_assignment *column_assign;
@@ -210,7 +210,7 @@ ir_mat_op_to_vec_visitor::do_mul_mat_scalar(ir_dereference *result,
 {
    int i;
 
-   for (i = 0; i < a->type->matrix_columns; i++) {
+   for (i = 0; i < (int)a->type->matrix_columns; i++) {
       ir_expression *column_expr;
       ir_assignment *column_assign;
