+++ 3rdparty/glsl-optimizer/src/glsl/lower_vec_index_to_swizzle.cpp
@@ -93,7 +93,7 @@ ir_vec_index_to_swizzle_visitor::convert_vec_index_to_swizzle(ir_rvalue *ir)
     * large.  For simplicity sake, just clamp the index to [0, size-1].
     */
    const int i = MIN2(MAX2(ir_constant->value.i[0], 0),
-		      (deref->array->type->vector_elements - 1));
+		      ((int)deref->array->type->vector_elements - 1));
 
    return new(ctx) ir_swizzle(deref->array, i, 0, 0, 0, 1);
 }
